% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtestV2.R
\name{backtestV2}
\alias{backtestV2}
\title{backtestV2 the given strategies & provide the 
  The difference between backtest & backtestV2 is , former technicallu loops all shorts/longs in one for loop. V2 - in different for loops.}
\usage{
backtestV2(pf, dataset, parms, startat = 2, longE, longX, shortE, shortX,
  calcLimits = calcLimits, ...)
}
\arguments{
\item{pf}{- portfolio to be tested}

\item{dataset}{- OHLCV with price objects, preferably xts}

\item{parms}{- parameters defining slp, profit booking etc. see tradeparms function}

\item{startat}{- From which row of the dataset to start testing. default 3}

\item{longE}{- A function which has conditions for long entry}

\item{longX}{- A function that has conditions for long exit}

\item{shortE}{- A function which has conditions for short entry}

\item{shortX}{- A function that has conditions for short exit}

\item{calcLimits}{- This is sample function given in this package, this can be overwritten by providing as parameter}

\item{...}{- Can pass instrument names like instr & other things needed by longE,shortE functions etc.}
}
\value{
portfolio name that has all the trades,transactions.
}
\description{
backtestV2 the given strategies & provide the 
  The difference between backtest & backtestV2 is , former technicallu loops all shorts/longs in one for loop. V2 - in different for loops.
}
\note{

}
