% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLimits.R
\name{calcLimits}
\alias{calcLimits}
\title{calcLimits for the given position
 Calculates the profitbooking price, pB slp, regular SLP, trailing target & trailing SLP etc}
\usage{
calcLimits(pos, tradeParms, m, i, ...)
}
\arguments{
\item{pos}{- current position}

\item{tradeParms}{- current tradeParms}

\item{m}{- dataset of the prices}

\item{i}{- the current bar in evaluation}
}
\value{
the position with updated prices
}
\description{
calcLimits for the given position
 Calculates the profitbooking price, pB slp, regular SLP, trailing target & trailing SLP etc
}
\note{

}
