% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{cBullishTrend}
\alias{cBullishTrend}
\title{All the constants required within the rulesNtrades package}
\format{An object of class \code{numeric} of length 1.}
\usage{
cBullishTrend
}
\description{
All the constants required within the rulesNtrades package
}
\keyword{datasets}
