% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawdowns.R
\name{drawdowns.trades}
\alias{drawdowns.trades}
\title{drawdowns - The sum of losing streaks in descending order}
\usage{
drawdowns.trades(t, n = 1)
}
\arguments{
\item{t}{- sequence of numbers to find drawdown}
}
\value{
numeric vector of drawdowns
}
\description{
- drawdown - is sum of sequence of all losing trades. netProfit is considered for calculating the same.
}
\details{
- This function gets the drawdowns in descending order
}

