% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph}
\alias{graph}
\title{graph - plot the graph into a given pdf}
\usage{
graph(pf, prices, file, by = "quarters", inTrades = NULL, overFun = NULL,
  ...)
}
\arguments{
\item{pf}{- portfolio which needs to be graphed}

\item{prices}{- OHLC of the shares/equity}

\item{file}{- filename where pdf to be stored. Please be advised, it will overwrite if any file exists.}

\item{by}{- frequncy for each page (whether monthly, weekly, quarterly) etc. 
e.g. monthly makes each page of pdf to contain one months data. 
All values accepted by endpoints function are available + halfyear. default - quarters.}

\item{inTrades}{- if given only these trades are graphed. portfolio is ignored.}

\item{overFun}{- overFun is a overlay function that helps to overlay graph wiht any of the indicators or values.}
}
\description{
This will plot the OHLC bars & corresponding trades performed.
}
\details{
Green uptriangle & Green downtriangle represents the Long Entry & Long Exits. Longs are drawn above the corresponding bar.

Red   uptriangle & red   downtriangle represents the short Entry & short Exits. Shorts are drawn below the corresponding bar.

overFun - is a function supplied by user. This is executed in the graph function environment, so all the variables
available for the graph function can be used by overFun. However caution to be exercised not to stop the function abruptly
}
\note{

}
\author{
Siva Sunku
}
\keyword{graph,plot}
