% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initObjects.R
\name{initObjects}
\alias{getTrxnFee}
\alias{initObjects}
\alias{setTrxnFee}
\title{Constructs the data containers used to store transactions and resulting trades}
\usage{
initObjects(n = 1000)

setTrxnFee(n = 0.1)

getTrxnFee()
}
\arguments{
\item{n-}{No. of rows to initialize trxns/trades. Default 1000}

\item{in}{- percentage of fee to be added for each transaction.}

\item{in}{- percentage of fee to be added for each transaction.}
}
\value{
nothing

trxnFee

trxnFee
}
\description{
initObjects will create all the required objects before starting the analysis

This will set the transaction fee provided in pct terms.
}
\note{
This will initialize all the objects required for the trade
}
\author{
Siva Sunku
}
\keyword{bars}

