% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{longEntrySample}
\alias{longEntrySample}
\title{The sample programs for longEntry,longExit,shortEntry,shortExit & calcLimits are given.
These can be copied and changed for specific logic.}
\usage{
longEntrySample(parms, m, i, ...)
}
\description{
The sample programs for longEntry,longExit,shortEntry,shortExit & calcLimits are given.
These can be copied and changed for specific logic.
}
