% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.blotter.R
\name{make.blotter}
\alias{make.blotter}
\title{make.blotter - Add all the transactions to a blotter package}
\usage{
make.blotter(in.port, out.port = "temp.blot", cur = "USD",
  symbol = "SBIN", initEq = 1e+05, initDate = "1950-12-31",
  verboseFlag = TRUE, forceClose = FALSE)
}
\arguments{
\item{in.port}{- rulesNtrades portfolio from which transactions to be obtained(input)}

\item{out.port}{- blotter portfolio to which rules to be added}

\item{cur}{- currency . Usually leave it as "USD"}

\item{symbol}{- stock which needs to be added (It won't filter the symbols from in.port. Create a different portfolio in rulesNtrades for each)}

\item{forceClose}{- Default FALSE. if already out.port is there, delete the portfolio. Remember all portfolios are deleted with this.}
}
\value{
portfolio name that has all the trades,transactions.
}
\description{
This will add the transactions in a given rulesNtrades portfolio to blotter package
 If you want to remove anything from blotter, use below commans
 rm(list=ls(envir=.blotter), envir = .blotter) 
 rm("portfolio.CUD",pos=.blotter)
 rm("account.CUD",pos=.blotter)  
 rm("order_book.CUD",pos=.strategy)
}
\author{
Siva Sunku
}
\keyword{graph,plot}
