% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{portfolio}
\alias{delete.portfolio}
\alias{is.portfolio}
\alias{is.valid.portfolio}
\alias{portfolio}
\alias{set.fee.portfolio}
\alias{update.portfolio}
\title{Create, Delete and Update a Portfolio}
\usage{
portfolio(pf = "default", n = 1000, Fee = 0.05)

is.portfolio(pf)

is.valid.portfolio(pf = "default")

delete.portfolio(pf = "default")

\method{update}{portfolio}(pf = "default", charexpr)

set.fee.portfolio(pf = "default", Fee = 0.05)
}
\arguments{
\item{n}{- number of trades in portfolio}

\item{Fee}{- Fees for each of the transaction in terms of percent. Default 0.05 (0.0005 X amount)}

\item{charexpr}{- character expression to be evaluated by update.portfolio}

\item{pf-}{name of the portfolio}
}
\value{
portfolio - returns name of the portfolio

is.portfolio - checks if the given name is a portfolio or not

is.valid.portfolio - Returns TRUE/FALSE

delete.portfolio - Returns the portfolio that is deleted or stops by error

update.portfolio - returns none or error from eval statement

set.fee.portfolio  - returns new transaction fee
}
\description{
initportfolio & getportfolio defaults the portfolio and retrieves the working environment.
}
\details{
portfolio - will initialize all the objects required for building the portfolio

is.portfolio - will return if a portfolio exists of that name & checks if given object is portfolio/not

is.valid.portfolio - will return if a portfolio exists of that name

delete.portfolio - will delete a given portfolio if it exists and is valid one

update.portfolio - will update any value inside a portfolio.

set.fee.portfolio - will set the transaction fee given in input. If 0.05 given, it is marked as 0.0005
}
\note{
This will have details of a portfolio
}
\author{
Siva Sunku
}
\keyword{portfolio}

