% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{position}
\alias{+.position}
\alias{-.position}
\alias{is.position}
\alias{isclosed}
\alias{isclosed.position}
\alias{islong}
\alias{islong.position}
\alias{isopen}
\alias{isopen.position}
\alias{isshort}
\alias{isshort.position}
\alias{position}
\alias{qty}
\alias{qty.position}
\title{position create an empty position Object}
\usage{
position(instr = "default")

is.position(pos)

isopen(x)

\method{isopen}{position}(pos)

isclosed(x)

\method{isclosed}{position}(pos)

islong(x)

\method{islong}{position}(pos)

isshort(x)

\method{isshort}{position}(pos)

qty(x)

\method{qty}{position}(pos)

\method{+}{position}(posA, posB)

\method{-}{position}(posA, posB)
}
\arguments{
\item{instr}{- instrument to be carried with position}
}
\value{
returns position object

returns TRUE if position is opened, FALSE otherwise
}
\description{
A position object will store the details of trade opening, price, qty, closing etc. 
Once a position is opened, till close the same object can be carried forward for further operations.

checks if given object is position type/class

check if position is open/closed
}
\details{
position - Creates an empty position object

isclosed - gives TRUE if qty is zero and position is opened

islong - checks if position is long/short

isshort - checks if position is long/short

qty - Gives the quantity of a position, according to the direction of the position. Returns +ve qty if long, -ve otherwise

"+" - Quantities are added & openprice is averaged

"-" - only quantities are substracted
}
\note{

}
\author{
Siva Sunku
}
\keyword{stats}

