% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions.R, R/trades.R
\name{position}
\alias{+.position}
\alias{-.position}
\alias{as.xts.trades}
\alias{isclosed.position}
\alias{islong.position}
\alias{isopen.position}
\alias{isshort.position}
\alias{position}
\alias{qty.position}
\title{positions}
\usage{
position(instr = "SHARE")

\method{isopen}{position}(pos)

\method{isclosed}{position}(pos)

\method{islong}{position}(pos)

\method{isshort}{position}(pos)

\method{qty}{position}(pos)

\method{+}{position}(a, b)

\method{-}{position}(a, b)

\method{as.xts}{trades}(t)
}
\arguments{
\item{instr}{- instrument that is being traded. Default "SHARE"}

\item{pos}{- position}

\item{a, b}{- two positions that needs to be added}

\item{t}{- table}

\item{pos}{- position}

\item{instr}{- instrument that is being traded.}

\item{instr}{- instrument that is being traded.}

\item{pos}{- position}

\item{a, b}{- two positions that needs to be added}
}
\value{
returns TRUE/FALSE

returns TRUE if position is opened, FALSE otherwise

returns TRUE if position is opened, FALSE otherwise

returns TRUE/FALSE

returns TRUE/FALSE

returns openqty if position is in open & long. returns -ve qty if position is short

returns new position which is sum of a and b

returns new position which is a - b

xts object of trades
}
\description{
Functions related to the positions

- The function averages the open price & adds the open Qty of both a & c. Remaining fields will be same as first object 'a;

- The function reduces the qty in b oject from a . Rest all fields will remain same
}
\details{
- This functions sums up two position objects

- This functions reduces the qty in b object from a
}
\note{
\code{position} returns empty position object that is not opened yet. position class object
}
\author{
Siva Sunku
}
\keyword{position}

