% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinvest.R
\name{reinvest.trades}
\alias{reinvest.trades}
\title{reinvest - reinvested trades}
\usage{
reinvest.trades(t, initCapital, stopCapital, buffCapital = 0, incrCapital,
  initQty, incrQty, maxQty = 10000)
}
\arguments{
\item{initCapital}{- Initial capital required}

\item{stopCapital}{- Capital threshhold, below which trading would be stopped & warning would be added}

\item{buffCapital}{- buffer capital, This would be saved before increasing the trade quantity}

\item{incrCapital}{- One Unit of incremental Capital, Each unit of incrCapital gives capacity to increase one unit of incrQty}

\item{initQty}{- Initial Quantity}

\item{incrQty}{- One unit of incremental quantity}

\item{maxQty}{- maximum quantity that can be traded. It will act as a roof irrespective of capital}
}
\value{
- Returns the new trades that would have been executed with incremental capital. New stats to be applied on this if stats to be observed.
}
\description{
This will provide the trades & their corresponding quantities, how they would be if all the capital was reinvested.
Capital gained can be utilized to increase the traded quantity according to the given trade parameters.
}
\note{

}
\author{
Siva Sunku
}
\keyword{graph,plot}
