% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{statObject}
\alias{statObject}
\alias{stats.trades}
\title{stats - Functions related to all stats - a trade is two transactions}
\usage{
statObject()

stats.trades(t)
}
\arguments{
\item{t}{- trades table}

\item{n}{- no. of rows}
}
\value{
trades dataframe

numeric vector of drawdowns
}
\description{
stats - Functions related to all stats - a trade is two transactions

- drawdown - is sum of sequence of all losing trades. netProfit is considered for calculating the same.
}
\details{
- This function returns a list of statistics object

- This function gets the statistics of trades
}
\note{

}
\author{
Siva Sunku
}
\keyword{position}

