% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeparms.R
\name{tradeParms}
\alias{tradeParms}
\title{Trading parameters}
\usage{
tradeParms(descr)
}
\value{
list object
}
\description{
All the trading parameters like slp, trailing slp etc. These are used in backtesting. Any parameters needed for execution
to be added and used in the functions.
}
\details{
- tradeparms - returns the trading parameters required
}
\note{

}
\author{
Siva Sunku
}
\keyword{parms}
