% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{add.trades.position}
\alias{add.trades.position}
\alias{get.trades}
\title{This function adds the position to a trade.}
\usage{
add.trades.position(pos)

get.trades(instr = NULL, dir = NULL)
}
\arguments{
\item{pos}{- position to be added}

\item{instr}{- default NULL, otherwise trades/trxns of particular instrument}
}
\value{
recently added trade

returns the trades/trxns data frame
}
\description{
This function adds the position to a trade.
}
\author{
Siva Sunku
}

