% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{zzz}
\alias{zzz}
\title{Onload package that gets executed whenever package is loaded.}
\usage{
zzz()
}
\description{
- Defines the objects gTrades,gTrxns,gStats 
the objects that are visible to this package


constants

Constants used within this package. These are available only in this package
}

