% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{backtest}
\alias{backtest}
\title{backtest the given strategies & provide the}
\usage{
backtest(pf, dataset, parms, startat = 2, longE, longX, shortE, shortX, ...)
}
\arguments{
\item{pf}{- portfolio to be tested}

\item{dataset}{- OHLCV with price objects, preferably xts}

\item{parms}{- parameters defining slp, profit booking etc. see tradeparms function}

\item{startat}{- From which row of the dataset to start testing. default 3}

\item{longE}{- A function which has conditions for long entry}

\item{longX}{- A function that has conditions for long exit}

\item{shortE}{- A function which has conditions for short entry}

\item{shortX}{- A function that has conditions for short exit}

\item{...}{- Can pass instrument names like instr & other things needed by longE,shortE functions etc.}
}
\value{
portfolio name that has all the trades,transactions.
}
\description{
All the rules that needs to be tested for evaluating a strategy. It initializes the trade table, trxn table.
Requires 4 functions as default for executing longEntry,longExit, shortEntry & shortExit trades.
Intra day conditions needs to be evaluated within the function code.
}
\note{

}
\author{
Siva Sunku
}
\keyword{backtest}

