% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{calcLimits}
\alias{calcLimits}
\title{Calculate the price limits like slp, trailing slp, etc}
\usage{
calcLimits(pos, tradeParms, bar, ...)
}
\description{
Calculate the price limits like slp, trailing slp, etc
}
