% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawdowns.R
\name{drawdowns}
\alias{drawdowns}
\title{Calculate the drawdowns in a series of profit}
\usage{
drawdowns(t, n = 1)
}
\arguments{
\item{t}{- sequence of numbers}

\item{n}{- max number of drawdowns, default - 1}
}
\value{
vector
}
\description{
drawdowns - Returns the drawdowns from highest losing streak to lowest losing streak. Sums the sequence of losses & returns
the same in descending order restricted by 'n' elements
}
