% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{statObject}
\alias{statObject}
\alias{stats.trades}
\alias{tabulate.trades}
\title{Statistics of all the trades}
\usage{
statObject()

stats.trades(t)

tabulate.trades(t)
}
\arguments{
\item{t}{- trades table. If t is provided, portfolio given will not be considered.}

\item{t}{- trades table. If t is provided, portfolio given will not be considered}
}
\value{
list object

numeric vector of drawdowns

Returns a list object that contains Description & xtab object next to it
}
\description{
All the statistics of trades executed in a given portfolio
}
\details{
- statObject - returns the statistics object

stats.trades - Calculates the statistics of a given trade table. 
If t is null, trades in portfolio are considered for calculation

tabulate.trades - Tabulates the trades according to year, month etc
If t is null, trades in portfolio are considered for calculation
}
\note{

}
\author{
Siva Sunku
}
\keyword{stats}

