% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{trades}
\alias{add.trades.position}
\alias{as.xts.trades}
\alias{get.trades}
\alias{is.trades}
\alias{trades}
\title{Trades in a portfolio}
\usage{
trades(n = 1)

is.trades(t)

add.trades.position(pf = "default", pos)

get.trades(pf = "default")

\method{as.xts}{trades}(t)
}
\arguments{
\item{n}{- no. of rows}

\item{t}{- trades}

\item{pf}{- portfolio name. default is "default"}

\item{pos}{- position to be added}

\item{pf}{- default "default"}
}
\value{
blank trades dataframe

recently added trade

returns the trades/trxns data frame

returns the trades/trxns data frame
}
\description{
A trade is two transactions a buy & sell transaction.

checks if given object is of trades class/not
}
\details{
trades - Returns the empty trades data frame with given 'n' objects

add.trades.position - Function to add given position to trades table in a given portfolio

get.trades - gets the trades from a default portfolio

as.xts.trades - Convert the trades from dataframe to xts. Since xts is a strict matrix, all string columns are removed.
Index of the returned xts object is open date on which trade is initially opened.
}
\note{
This will create a trade object
}
\author{
Siva Sunku
}
\keyword{trades}

