% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trxns.R
\name{trxns}
\alias{add.trxns.position}
\alias{get.trxns}
\alias{is.trxns}
\alias{trxns}
\title{Transactions(trxns)}
\usage{
trxns(n = 1)

is.trxns(t)

add.trxns.position(pf = "default", pos, type = c("OPEN", "CLOSE"))

get.trxns(pf = "default")
}
\arguments{
\item{n}{- no. of rows of blank transaction dataframe.}

\item{pf}{- portfolio where the transactions to be added. Default - default portfolio}

\item{pos}{- position to be added}

\item{type}{- OPEN - transaction is added for opening position, CLOSE transaction is added while closing position}

\item{pf}{- default is "default" portfolio}
}
\value{
recently added transaction

returns the trades/trxns data frame
}
\description{
A transaction is a buy/sell of a position. Two transactions make a trade(buy/sell).

checks if given object is trxns class/not
}
\details{
- This function returns the dataframe of 'n' blank transactions.

add.trxns.position - adds a given position to the transaction table of a given portfolio
}
\note{

}
\author{
Siva Sunku
}
\keyword{transactions}

